JsonWriter: refactored to generalize handling of different cases
In order to reduce the code size, and the number of cases handled,
refactored JsonWriter to use the same approach as in the PrettyWriter,
with pattern-matching against operators in the map.

It is unclear what names to employ for TLA operators in JSON encoding.
Currently, the most reasonable approach seems to use ASCII names
of TLA operators with several exceptions to avoid ambiguity.