Modified Dockerfile to check whether a user dir was bind-mounted
With the previous version of Dockerfile it was not possible to access
specification files from inside the container. For that a user directory
needs to be bind-mounted into the container using docker -v option.

Added a simple ENTRYPOINT script, bin/run-in-docker-container, that checks that
some .tla files are present in /var/apalache, i.e. it was indeed bind-mounted.