make --init and --next checker options detectably optional
Until now it was not possible to detect whether the user provided
init/next, or just the default values were used.

As with the parsing of TLC config there are several possible
sources for init/next, we need to detect and report conflicts,
when they are provided in multiple places.