ConfigRewriter: preliminary topological sort of TLA definitions
Due to the way how TLA Toolbox introduces constants and initializes them
in TLC config, the constants get initialized _after_ they are used.

As Apalache replaces constants from a TLC config inline, this creates
a problem of a value being used before being defined.

To mitigate this problem, we do a topological sort of TLA definitions,
according to their dependency graph.

This commit is a preliminary (incomplete) implementation of it.