Rework the introduction of type annotation syntax
It took my some time to figure out what was being explained here. I
finally realized my confusion was owed in part to the fact that the
presentation simultaneously introduces the type ascription operator and
what is essentially a hack to compel other tools to ignore the operator.

I've reworded the section to try to make these two different aspects
clear. This change also suggests renaming the variables used to clearly
indicate the value and the type.