Expand expanation of record subtyping
I found the account of record types pretty unsatisfactory. In
particular, the reference to C unions seemed like a misdirect to me. I
suspect that other readers will find it helpful to read a more explicit
account of the subtyping relationship that makes this type checking
possible.

I'm not at all sure if I'm using the preferred notation or terminology
when describing the types of TLA records, or if the reference to row
polymorphism is helpful (tho I do think it is apt).