Tweaks to manual and addition of CONTRIBUTING.md (#155)
* Tweak phrasing and punctuation

* Rework the introduction of type annotation syntax

It took my some time to figure out what was being explained here. I
finally realized my confusion was owed in part to the fact that the
presentation simultaneously introduces the type ascription operator and
what is essentially a hack to compel other tools to ignore the operator.

I've reworded the section to try to make these two different aspects
clear. This change also suggests renaming the variables used to clearly
indicate the value and the type.

* Consistently place anchors directly over headings

Some section anchors were immediately above the corresponding heading
and some had a blank space between. In my effort to make it uniform, I
opted for the formatting I prefer, but I'm happy to make it uniform the
other way instead :)

* Slight rephrase and reference to CONTRIBUTING.md

* Add a direnv rc file and document its (optional) use

direnv is a super helpful too for automating project-specific shell
environment configuration. This change adds an rc exporting the relevant
env vars, and putting the project's `bin` dir in the uses path (but only
when they are using direnv and working out of the the project's
repository).

* Reorder installation instructions

This orders the sequence of installation steps so that all the system
dependencies are installed first, allowing us to give clearer and more
specific instructions about the stuff specific to installing apalache.

This also cleans up the links a bit and expands on a few points.

* Show example of failing counterexample

* Add CONTRIBUTING.md

This is taken largely from the example in the handbook, with some tweaks
for our project, and some requirements that don't seem relevant to
us (yet) removed.

* Add top-level heading to collection sections on Apalache principles

This change groups three top-level sections under a new heading

1. [Principles of symbolic model checking with Apalache](#principles)
   1. [Assignments and symbolic transitions](#assignments)
   1. [Type annotations](#types)
   1. [Recursive operators and functions](#recursion)

* Add Make target for unit tests

Also declares the phony targets to be so.