Update dependencies
Base update achieved via mvn versions:use-latest-releases, then
moderated by decrementing from versions marked `-alpha` or `-beta`.

This also postpones updating the testing library, because it broke the
tests (all approach this in a followup), and adds dependency
speciicaitons for the logging backend. The need and explanation for the
later is documented in a comment.