Make Dockerfile staged (#195)
* Make Dockerfile staged

This builds the vendored dependencies and apalache itself in three
separate stages. This lets us reuse previously built images for
z3 and box, so long as the content of the ./3rdparty direcotry
hasn't been changed, and puts the resulting executables into a minimal
app image. As a result, final size of the docker container is reduced by 
nearly 1GB

* Remove unneeded surefire workaround in docker recipe

Checked this by running the integration tests successfully with the
resulting image built

* Add script to wrap docker command

* Add flag to run integration tests using a docker image

This makes it easy to run the integration tests against a docker
image.