Simplify record validation
Use simple structural matching instead of arithmetic. This change
also lets us validate the args in a single pass, rather than two
separate invocations of `zipWithIndex`.

It actually increases the loc, but that's due to more robust error
handling and the addition of unit tests to cover an unhappy path.