Refactor mkApp helper
This renames the helper so that it doesn't shadow the original function
(this shadowing was leading to confusion for this reader :)), and it
makes the helper a closure that includes the exact ref id, so that
needn't be supplied every time.

It also adds a temporary variable so we only need to construct the
ExactRef once.