Compile the manual into an mdbook-gemerated static site (#400)
* Remove the version number in the text body

This should be derived from the source code. The fact that we keep the
manual under the same version control as the source ensures that they
stay in sync, we don't need manually text notes in the body (which we
will inevitably forget to update! :))

* Move the intro paragraph into its own chapter

This is important for the way that mdbook works.

* Add mdbook SUMMARY.md

* Move all manual pages into docs/src

This provides better organization, now that we're compiling the source
files into a book, and follows mdbook conventions.

* Add note on SUMMARY.md

* Update manual location

We will replace these shortly links with links to the hosted HTML
compiled by mdbook, but for this PR I'm just keeping them update with
the new locations in the src.

* Add CI to build and deploy the docs

* Remove the .nojekyll file

This breaks the jekyll rendering from the github servers

* Ignore website artifacts

* Remove manual "Back to top" links

These aren't necessary now that we're compiling to mdbook.

* Update book config

- Don't create empty files for toc entries without pages
- Enable mathjax output
- Link to git repository

* Move "hairy" docs into internal directory

At Igor's request

* Fix links to the manual throughout the codebase

Point them to the compiled, hosted manual, instead of the source
markdown files.

* added a section on running apalache in java + fix by Leslie