Fix the release PR title
Follow up to #612

The wrong phrase ended up as the title of release PRs. The first
`--message` argument given to `hub` will end up as the title of the PR,
which should be the commit. This is change returns the commit message of
the release commit to that position, as intended.