Add CI set up for smt-encoding work (#1061)
* Simplify run.txt generation

This removes the brittle, and hardcoded report on CLI args,
instead just threading through the information that is already available
in from the arg array when the commands are parsed.

* Remove the run-integration script

That wrapper has become useless

* Get SMT_ENCODING envvar via CLI configuration

Apparently clist has a mechanism for deriving the envvar
value automatically. We should use that.

* Print the output directory

This makes it much easier to figure out what the output has gone, cause
you don't have to parse through the long, dense file names.

* Record env-vars gathered by CLI in run.txt

* Add integration test for reporting env vars set by CLI library

* Add smt-encoding configuration to CI matrix

* Add `array-encoding` TEST_FILTER

This allows filtering in just those integration tests we want to run
during development of the experimental array smt encoding.