Sketch of a type-level expression build
Closes #1102

This presents a very small (very rough) example of how we might push
guaranteed type-correct TLA expressions into the type level. It is meant
to demo the general technique, so we can get a rough sense of whether
it's worth pursuing.

I suspect there may be nonobvious complications that would arise in the
process of trying to apply such an approach, and porting the existing
code would involve a lot of work.

The benefit to hope for is that we might eliminate large sections of
code and supporting tests.