Move all RFCs into same directory
Keeping our rfcs/adrs in separate directories causes me confusion, and I
imagine would do so for others too. Moreover, I think it's not useful.
When we initially separated them out, it was to protect users from
encountering design documents that wouldn't be useful to them. Now we
can do this by just omitting the docs we wish to hide from the listing
in `SUMMARY.md`.