Add integration tests for configuration system
We just check overrides for one config value, since we aren't checking
that the PureConfig library can load configs, only that our desired
precedence levels and overrides work as expected via the PureConfig
library.

I've tested the default global config locally, as I'm not sure of a good
way to put that in the integration tests.