Silence package data output by default
Supports #1278

Currently, the apalache-mc script produces 4 lines of output on every
run, printout where the script is being run from and where the jar is
located.

This info is generally not useful, so this change hides that output
behind an environment variable you can set if you want to get that
output printed. This can be useful for debugging.