Use exec to run java
This ensures that the process of the runner script is replaced by the
launched JVM process, rather than running it in a subshell, this should
give us the same benefit of sigterm propagation without needing a trap.
This approach is adapted from the runner script that is generated by
sbt-native-packager when building java-app archetypes (we might consider
further adapting our packaging to use in the future).