Add auto-update github action
This should help alleviate the tedium resulting from the PR
updating/merging race condition.

When changes are pushed to our default branch, this action is meant to
check any PRs that have 'auto-merge' enabled, and trigger their update.

If everything works as intended, then once a PR is ready for merge and
we've enabled `auto-merge`, this action will do the work of ensuring the
PR is kept up to date in case a different PR is merged in before the CI
can complete.