Add ADR-21: Prioritization of work
Closes #936

This proposes a process for deciding, communicating, and recording the
prioritization of our work. Please see the ADR for more information on
context and motivation.

Note that this is just a proposal to get the discussion going, and we
should make any changes we see fit to adapt the proposal to our needs.
Even if, on review, we end up deciding we don't want the extra overhead
of this kind of process, we'll at least have made a clear stand on that
:)