Fix netty packaging issue
This was giving a an error
`deduplicate: different file contents found in the following` due
to certain duplicate META-INF resources during packaging. This is
because netty is already pulled in as a dependency, but still need
to define it for use within our packages. Thus the "provided" scope.

See https://stackoverflow.com/a/26157704/1187277