Encapsulate execution logic of each subcommand
Each subcommand provides an API for configuring a different kind of
process or sub-program. Prior to this change, all of the top-level of
these sub-programs was intermixed in the `Tool.scala` module. This
change factors out the specific logic of each subcommand's configuration
and execution into a `run` method in the class defining the subcommand.
This follows the idiom of "adding behaviors to commands" recommended by
clist: https://github.com/backuity/clist#adding-behavior-to-commands