Refactor SanyParserPassImpl
A light refactor of the `execute` method in the `SanyParserPassImpl`.
I prepared this initially to support #1114, but realized it wasn't quite
the right way to meet the needs of the server.

Still, I think it's a nice enough cleanup to be worth integrating.

The refactor just breaks the execution logic into three chunks with some
helper functions and uses a for comprehension over `Either` to sequence
the possible error cases.