Replace `options.filename` with `options.source`
The essential change here is to replace the "filename" option, storing a
string, with an instance of the newly added `SourceOption` class, that
lets us statically encode a data source which is (currently] either a
file or a string.

The immediate use for this will be supporting Shai to load modules from
a string (while maximizing code reuse), but this could also be used
to read in specs (or other data) from stdin or a socket etc.