Fix scaladoc errors
This attempts to fix all the scaladoc reference errors encountered while
building the scaladoc for our codebase. There were numerous cases were I
could not figure out any way to point scaladoc to the intended value,
even when it seems like it should have been in scope. But most changes
are due to incorrectly attempting to refer to private values or not
qualifying references correctly etc.