Add static representation of configuration options (#2024)
Contributes towards #1174

This adds traits with values to represent all the options that we are
currently setting in the `at.forsyte.apalache.tla.tooling.opt.Cmd*`
classes. None of those options are actually transferred over here.
Instead, this changeset just prepares the representation of the options,
with the aim of obtaining feedback on the presentation.