Make commands invoke ApalacheCommand
This moves the invocation of the `ConfigManager` inside of the
`ApalacheCommand` class. The result is that all needed application
configuration coming from the outside world is now available from a
successfully constructed instance of `ApalacheCommand`. This will become
very useful in followup changes that start migrating from the untyped
options map to objects satisfying the options traits, since this will
enable every process executed by the `cmd.run` method to have access to
the loaded application configuration.