Add `routine` configuration
This tracks what "routine" of the program is being run. This gives us a
place to pass around information about which commands or services are
being run. The main motivating force here is so that we can drop the
dummy file value we're producing for subcommands that don't actually
rely on files.

These dummy files were previously used by output manager to derive names
for the output files.