Expose the JSON representation of TlaModules
In order to incorporate the JSON representation of a TLA module into
replies sent to RPC requests, we need to embed the JSON represention
in other JSON data structures. The current mechanism for generating JSON
of a `TlaModule` only allowed serializing to a string, but didn't
provide a mechanism for obtaining the JSON representation.  This change
exposes the underlying representation.