Consolidate RPC Server test suite
We will want to organize the tests for the RPC server into several
different modules, e.g. to group the different services or different RPC
calls into different files. However, given the current architecture, we
need all tests of the RPC server to share the same semaphore to protect
access to SANY.

This change separates out the definition of test cases from the
definition of the suite which actually runs the test cases. Thus, we are
able to define the tests together merely as a sequence, then run all the
tests together with the same ZIO service layers, thereby ensuring they
have the same semaphore.