Adapt tests to newline preserving annotations
The tests and implementation of reprocessing and parsing of comments and
annotations is pretty brittle. It is sensitive to white space in ways
that are (imo) fairly surprising. I've needed a few dozen little tweaks
and fixes to the parser and test in order to adapt them  to the fix in
the annotation parser.