Fix unhandled exception on invalid file formats
Closes #2175

This fixes the processing of CLI configuration params to avoid unhandled
exceptions. The solution is to be honest about the possible failure of
the the `toConfig` method, which maps CLI param inputs to an
instance of `Options.ApalacheConfig`. Since this conversion can fail, it
should be a `Try`. This change fixes the type.

Along the way, it removes the atavistic `collectTuningOptions` method,
which shouldn't be needed with the new config management scheme. This
also enables a slight cleanup on the Tracee command.