Refactor quint expression conversion tests
It's becoming difficult to manage the type mapping for the
quint conversion tests, because currently we have to do double entry
when constructing the expressions and their types.

This ugly but pragmatic refactor makes it so that the construction of
expressions and registering their types happens in one go.

This PR is mostly mechanical, and is factored out of the fix for #2516.