Generalize NullaryOpReader into ContextReader
We need to track both the names of nullary operators in scope and whether or not
we in are in the scope of the `q::init` operator. The latter is require
so that we can ensure all assignments in the init operator are unprimed.
So we expand the reader to store both these bits of data